rotc_backend/requirements.txt
# Core Django
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0

# Database
psycopg2-binary>=2.9,<3.0
dj-database-url>=2.1,<3.0

# CORS
django-cors-headers>=4.3,<5.0

# Environment variables
python-dotenv>=1.0,<2.0

# File storage and processing
django-storages>=1.14,<2.0
cloudinary>=1.36,<2.0
Pillow>=10.0,<11.0

# Real-time and background processing
celery>=5.3,<6.0
redis>=5.0,<6.0
django-redis>=5.4,<6.0
django-celery-results>=2.5,<3.0
django-celery-beat>=2.5,<3.0
channels>=4.0,<5.0
channels-redis>=4.1,<5.0
daphne>=4.0,<5.0

# Document processing
ReportLab>=4.0,<5.0
pytesseract>=0.3,<1.0

# Push notifications
django-push-notifications>=3.0,<4.0

# Password hashing
bcrypt>=4.1,<5.0

# Rate limiting
django-ratelimit>=4.1,<5.0

# Production server
gunicorn>=21.2,<22.0

# Utilities
openpyxl>=3.1,<4.0
prometheus-client>=0.19,<1.0

# Error tracking (optional)
sentry-sdk>=1.40,<2.0
